pyproject.toml
[tool.poetry]
name = "py-dagger-contrib"
packages = [{ include = "dagger_contrib" }]
description = "Extensions for the Dagger library (py-dagger in PyPI)."
authors = ["larribas <lorenzo.s.arribas@gmail.com>"]
version = "0.0.0"
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/larribas/dagger-contrib"
keywords = [
  "dagger",
  "py-dagger",
  "data-engineering",
  "data-pipelines",
  "data-science"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Build Tools"
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
py-dagger = "^0.3.0"
PyYAML = { version = "^5.4", optional = true }
pandas = { version = "^1.3", optional = true }

[tool.poetry.extras]
yaml = ["PyYAML"]
pandas = ["pandas"]
all = ["PyYAML", "pandas"]

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^20.8b1"
isort = "^5.7.0"
flake8 = "^3.9.2"
pydocstyle = "^6.1.1"
mypy = "^0.812"
pytest-cov = "^2.12.0"
pyarrow = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pydocstyle]
inherit = false
convention = "numpy"

